L3: Centralize PWA and service worker mocks - reduce duplication
### Issue
- PWA navigator.serviceWorker and sync manager mocks defined inline in utils.unit.test.ts
- Scattered utilities (buildSW, buildPerm, mockWorker, restorePwaGlobals) only in one file
- Difficult to reuse or extend for new PWA tests

### Fix
- Created shared pwaTestHelper.ts with centralized PWA mock utilities
- Exported: mockPwaWorker, buildSW, buildPerm, restorePwaGlobals, PwaTestState type
- Updated utils.unit.test.ts to import from shared helper
- Replaced all mockWorker() calls with mockPwaWorker()

### Benefits
- Reusable PWA mocks available for other test files
- Consistent interface across tests
- Easier to maintain and extend PWA mock behavior
- Clear separation of concerns (test helpers vs test code)

### Verification
- All 1421 tests passing
- PWA utils tests cover sync/periodicSync/permissions scenarios
- No regressions
