+++ frontend/src/config/pwa/utils.unit.test.ts
@@ -1,87 +1,16 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { describe, it, expect, vi, afterEach } from 'vitest';
+import { describe, it, expect, afterEach } from 'vitest';
 
 import { registerBackgroundSync, registerPeriodicSync } from './utils';
-
-const originalNavigator = navigator;
-const originalWindow = globalThis;
-
-type TestState = 'have' | 'donthave' | 'fail';
-
-const delayPromiseRejection = (ms: number, error: any) =>
-  vi.fn(() => new Promise((_, reject) => setTimeout(() => reject(error), ms)));
-
-const buildSW = (fail: boolean) => {
-  if (fail) {
-    return {
-      serviceWorker: {
-        ready: delayPromiseRejection(100, 'Service Worker error'),
-      },
-    };
-  }
-  return {
-    serviceWorker: {
-      ready: Promise.resolve({
-        periodicSync: {
-          register: vi.fn(),
-        },
-        sync: {
-          register: vi.fn(),
-        },
-      }),
-    },
-  };
-};
-
-const buildPerm = (testState: TestState, state: PermissionState) => {
-  switch (testState) {
-    case 'have': {
-      return {
-        permissions: {
-          query: vi.fn().mockResolvedValueOnce({ state }),
-        },
-      };
-    }
-    case 'donthave': {
-      return {};
-    }
-    case 'fail': {
-      return {
-        permissions: {
-          query: delayPromiseRejection(100, 'Permission error'),
-        },
-      };
-    }
-  }
-};
-
-const mockWorker = (
-  state: PermissionState,
-  periodicSync: TestState,
-  sync: TestState,
-  permission: TestState,
-) => {
-  vi.stubGlobal('navigator', {
-    ...buildSW(periodicSync === 'fail' || sync === 'fail'),
-    ...buildPerm(permission, state),
-  });
-  // Provide stubbed managers when feature is expected to exist
-  if (periodicSync === 'have')
-    (globalThis as any).PeriodicSyncManager = (globalThis as any).PeriodicSyncManager || {};
-  if (sync === 'have') (globalThis as any).SyncManager = (globalThis as any).SyncManager || {};
-  if (periodicSync === 'donthave') delete (globalThis as any).PeriodicSyncManager;
-  if (sync === 'donthave') delete (globalThis as any).SyncManager;
-};
+import { mockPwaWorker, restorePwaGlobals } from '@/config/tests/pwaTestHelper';
 
 afterEach(() => {
-  (globalThis as any).PeriodicSyncManager = (originalWindow as any).PeriodicSyncManager;
-  (globalThis as any).SyncManager = (originalWindow as any).SyncManager;
-  vi.stubGlobal('navigator', originalNavigator);
+  restorePwaGlobals();
 });
 
 describe('registerBackgroundSync', () => {
   it('should register background sync when supported and granted', async () => {
-    mockWorker('granted', 'have', 'have', 'have');
+    mockPwaWorker('granted', 'have', 'have', 'have');
 
     const result = await registerBackgroundSync('test-tag');
     expect(result).toBe(true);
@@ -91,21 +20,21 @@ describe('registerBackgroundSync', () => {
   });
 
   it('should return false if permission is denied', async () => {
-    mockWorker('denied', 'have', 'have', 'have');
+    mockPwaWorker('denied', 'have', 'have', 'have');
 
     const result = await registerBackgroundSync('test-tag');
     expect(result).toBe(false);
   });
 
   it('should return false if SyncManager is not supported', async () => {
-    mockWorker('granted', 'have', 'donthave', 'have');
+    mockPwaWorker('granted', 'have', 'donthave', 'have');
 
     const result = await registerBackgroundSync('test-tag');
     expect(result).toBe(false);
   });
 
   it('should return false ready failed', async () => {
-    mockWorker('granted', 'have', 'fail', 'have');
+    mockPwaWorker('granted', 'have', 'fail', 'have');
 
     const result = await registerBackgroundSync('test-tag');
     expect(result).toBe(false);
@@ -114,7 +43,7 @@ describe('registerBackgroundSync', () => {
 
 describe('registerPeriodicSync', () => {
   it('should register periodic sync when supported and granted', async () => {
-    mockWorker('granted', 'have', 'have', 'have');
+    mockPwaWorker('granted', 'have', 'have', 'have');
 
     const result = await registerPeriodicSync('periodic-tag', 3600000);
     expect(result).toBe(true);
@@ -124,31 +53,31 @@ describe('registerPeriodicSync', () => {
   });
 
   it('should return false if PeriodicSyncManager is not supported', async () => {
-    mockWorker('granted', 'donthave', 'have', 'have');
+    mockPwaWorker('granted', 'donthave', 'have', 'have');
     const result = await registerPeriodicSync('periodic-tag', 1000);
     expect(result).toBe(false);
   });
 
   it('should return false if permission is denied', async () => {
-    mockWorker('denied', 'have', 'have', 'have');
+    mockPwaWorker('denied', 'have', 'have', 'have');
     const result = await registerPeriodicSync('periodic-tag', 1000);
     expect(result).toBe(false);
   });
 
   it('should return false if fail to get ready', async () => {
-    mockWorker('granted', 'fail', 'have', 'have');
+    mockPwaWorker('granted', 'fail', 'have', 'have');
     const result = await registerPeriodicSync('periodic-tag', 1000);
     expect(result).toBe(false);
   });
 
   it('should return false if permission fail', async () => {
-    mockWorker('granted', 'fail', 'have', 'fail');
+    mockPwaWorker('granted', 'fail', 'have', 'fail');
     const result = await registerPeriodicSync('periodic-tag', 1000);
     expect(result).toBe(false);
   });
 
   it('should return false if permission dont exist', async () => {
-    mockWorker('granted', 'fail', 'have', 'donthave');
+    mockPwaWorker('granted', 'fail', 'have', 'donthave');
     const result = await registerPeriodicSync('periodic-tag', 1000);
     expect(result).toBe(false);
   });
